+++ packages/react-kit/src/components/DefaultTheme/theme.ts
@@ -11,7 +11,7 @@ import { CHECKBOX } from '../Checkbox/Checkbox';
 import * as checkboxTheme from '../Checkbox/Checkbox.demo.styl';
 
 import { POPOVER } from '../Popover/Popover';
-import * as popoverTheme from '../Popover/Popover.demo.styl';
+import * as popoverTheme from '../Popover/demo/theme/Popover.demo.styl';
 
 import { LIST } from '../List/List';
 import * as listTheme from '../List/List.demo.styl';
@@ -58,7 +58,7 @@ import * as popupTheme from '../Popup/Popup.demo.styl';
 import { STEPPABLE_INPUT } from '../SteppableInput/SteppableInput';
 import * as steppableInput from '../SteppableInput/theme/SteppableInput.styl';
 import * as steppableInputInput from '../SteppableInput/theme/Input.styl';
-import * as steppableInputButtonIcon from '../SteppableInput/theme/ButtonIcon.styl';
+import * as steppableInputButtonIcon from '../SteppableInput/theme/ButtonIcon.demo.styl';
 
 import { TIME_INPUT } from '../TimeInput/TimeInput';
 import * as timeInput from '../TimeInput/theme/TimeInput.styl';