+++ packages/react-kit/src/components/Deferred/Deferred.page.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { PureComponent } from 'react';
-import { storiesOf } from '@devexperts/tools/dist/utils/storybook';
+import { storiesOf } from '@storybook/react';
 import Demo from '../demo/Demo';
 import { Deferred } from './Deferred';