+++ packages/react-kit/src/components/ResizeDetector/ResizeDetector.page.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import { storiesOf, action } from '@devexperts/tools/dist/utils/storybook';
+import { storiesOf } from '@storybook/react';
+import { action } from '@storybook/addon-actions';
 import { ResizeDetector } from './ResizeDetector';
 import Demo from '../demo/Demo';