+++ packages/react-kit/src/components/SteppableInput/SteppableInput.tsx
@@ -21,6 +21,8 @@ export type TPickedInputProps = Pick<
 
 export type TFullSteppableInputProps = TPickedInputProps & {
 	isDisabled?: TInputProps['isDisabled'];
+	isIncrementButtonDisabled?: boolean;
+	isDecrementButtonDisabled?: boolean;
 	tabIndex?: number;
 	onIncrement?: Function;
 	onDecrement?: Function;
@@ -78,6 +80,8 @@ class RawSteppableInput extends React.Component<TFullSteppableInputProps, TStepp
 		} = this.props;
 
 		const { isFocused } = this.state;
+		const isIncrementButtonDisabled = isDisabled || this.props.isIncrementButtonDisabled;
+		const isDecrementButtonDisabled = isDisabled || this.props.isDecrementButtonDisabled;
 
 		return (
 			<Input
@@ -112,26 +116,26 @@ class RawSteppableInput extends React.Component<TFullSteppableInputProps, TStepp
 						)}
 					{onDecrement &&
 						decrementIcon && (
-							<Holdable onHold={onDecrement}>
+							<Holdable onHold={onDecrement} isDisabled={isDecrementButtonDisabled}>
 								<ButtonIcon
 									icon={decrementIcon}
 									theme={theme.ButtonIcon}
 									onClick={this.onDecrementClick}
 									onMouseDown={this.onButtonMouseDown}
-									isDisabled={isDisabled}
+									isDisabled={isDecrementButtonDisabled}
 									tabIndex={-1}
 								/>
 							</Holdable>
 						)}
 					{onIncrement &&
 						incrementIcon && (
-							<Holdable onHold={onIncrement}>
+							<Holdable onHold={onIncrement} isDisabled={isIncrementButtonDisabled}>
 								<ButtonIcon
 									icon={incrementIcon}
 									theme={theme.ButtonIcon}
 									onClick={this.onIncrementClick}
 									onMouseDown={this.onButtonMouseDown}
-									isDisabled={isDisabled}
+									isDisabled={isIncrementButtonDisabled}
 									tabIndex={-1}
 								/>
 							</Holdable>