+++ packages/react-kit/src/components/input/Input.page.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import { storiesOf, action } from '@devexperts/tools/dist/utils/storybook';
+import { storiesOf } from '@storybook/react';
+import { action } from '@storybook/addon-actions';
 import { Demo } from '../demo/Demo';
 
 import { Input } from './Input';