+++ packages/react-kit/src/utils/react.ts
@@ -1,3 +1,2 @@
-export * from './css';
 export * from './pure';
 export * from './disposable';