+++ packages/rx-utils/src/observable-map.utils.ts
@@ -26,7 +26,8 @@ export class ObservableMap<K, V> {
 	private isInTransaction = false;
 	private hasChanges = false;
 
-	readonly keys$: Observable<K[]> = this.allSubject$.pipe(map(() => Array.from(this.cache.keys())), shareReplay(1));
+	private _keys$ = new BehaviorSubject<K[]>([]);
+	readonly keys$ = this._keys$.asObservable();
 
 	readonly values$: Observable<V[]> = this.allSubject$.pipe(
 		map(() => {
@@ -51,6 +52,8 @@ export class ObservableMap<K, V> {
 		return this.size === 0;
 	}
 
+	private handleKeys = (keys: K[]) => this._keys$.next(keys);
+
 	has(key: K): boolean {
 		return this.cache.has(key);
 	}
@@ -70,10 +73,14 @@ export class ObservableMap<K, V> {
 
 	set(key: K, value: V): void {
 		this.transaction(() => {
+			const isCachedKey = this.cache.has(key);
 			let cached = this.getOrCreateCached(key);
 			if (cached.hasValue === false) {
 				cached = initializeEntity(cached);
 				this.cache.set(key, cached);
+				if (!isCachedKey) {
+					this.handleKeys(Array.from(this.cache.keys()));
+				}
 			}
 
 			if (cached.subject.getValue() !== value) {
@@ -109,7 +116,11 @@ export class ObservableMap<K, V> {
 
 	delete(key: K): void {
 		this.transaction(() => {
+			const isCachedKey = this.cache.has(key);
 			this.cache.delete(key);
+			if (isCachedKey) {
+				this.handleKeys(Array.from(this.cache.keys()));
+			}
 			this.hasChanges = true;
 		});
 	}
@@ -125,6 +136,7 @@ export class ObservableMap<K, V> {
 	clear(): void {
 		this.transaction(() => {
 			this.cache.clear();
+			this.handleKeys(Array.from(this.cache.keys()));
 			this.hasChanges = true;
 		});
 	}