+++ packages/rx-utils/CHANGELOG.md
@@ -3,6 +3,51 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.29.0](https://github.com/devex-web-frontend/dx-platform/compare/v0.28.3...v0.29.0) (2019-05-25)
+
+**Note:** Version bump only for package @devexperts/rx-utils
+
+
+
+
+
+
+## [0.28.3](https://github.com/devex-web-frontend/dx-platform/compare/v0.28.2...v0.28.3) (2019-04-22)
+
+**Note:** Version bump only for package @devexperts/rx-utils
+
+
+
+
+
+## [0.28.2](https://github.com/devex-web-frontend/dx-platform/compare/v0.28.1...v0.28.2) (2019-04-16)
+
+
+### Bug Fixes
+
+* **rx-utils:** keys$ updates if keys number has changed ([#111](https://github.com/devex-web-frontend/dx-platform/issues/111)) ([2e6c831](https://github.com/devex-web-frontend/dx-platform/commit/2e6c831))
+
+
+
+
+
+## [0.28.1](https://github.com/devex-web-frontend/dx-platform/compare/v0.28.0...v0.28.1) (2019-04-11)
+
+**Note:** Version bump only for package @devexperts/rx-utils
+
+
+
+
+
+
+# [0.28.0](https://github.com/devex-web-frontend/dx-platform/compare/v0.27.0...v0.28.0) (2019-03-22)
+
+**Note:** Version bump only for package @devexperts/rx-utils
+
+
+
+
+
 # [0.27.0](https://github.com/devex-web-frontend/dx-platform/compare/v0.26.0...v0.27.0) (2019-03-12)
 
 **Note:** Version bump only for package @devexperts/rx-utils